+++ crates/web-client/src/models/sparse_merkle_path.rs
@@ -26,7 +26,7 @@ impl SparseMerklePath {
 
     /// Verifies the path against a root.
     pub fn verify(&self, index: JsU64, node: &Word, root: &Word) -> bool {
-        self.0.verify(index as u64, node.clone().into(), &root.clone().into()).is_ok()
+        self.0.verify(index, node.clone().into(), &root.clone().into()).is_ok()
     }
 }
